pyproject.toml
[project]
name = "daily-report-generator"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "kai", email = "-"},
]
dependencies = ["pillow>=11.1.0", "pandas>=2.2.3", "openpyxl>=3.1.5", "python-docx>=1.1.2"]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
